enable saving cosmetics to cookie for re-use between seeds
Adds ability to push button under Cosmetics Options to save cosmetics
settings to a cookie ("ctjot_cosmetics"). When that cookie is present,
when seed generated, values are read from the cookie and pre-populate
settings on the seed share page. Also adds ability to explicitly re-load
from cookie and reset to defaults (which also removes cookie).