Fix first time seed hash error
On the first generation of a seed hash it would be treated as a
different data type than if the seed hash is pulled from the DB.
This was causing type conversion issues while decoding the symbols
for the image.  Change indexing method to force individual values
to be treated as bytes, rather than sometimes being bytes and
sometimes being integers.