use MD5 sums of static js files to cache-bust
Consistently get browser to reload on changes to static .js files
by computing MD5 checksums at each during django start (in AppConfig),
injecting into views, and passing as a "?v" query for request to get
script. This would only be for in dev or when DEBUG is set because
otherwise django's ManifestStaticFilesStorage handles this.