enable unlocked skyways when rocksanity; revert to prior value
When Rocksanity is toggled on, turn on Unlocked Skyways. When
Rocksanity is toggled off, revert Unlocked Skyways to prior state
before Rocksanity was toggled on (i.e. keep on if already on before
Rocksanity was toggled or turn off if it was off before Rocksanity
was toggled). Disable Rocksanity if Unlocked Skyways gets disabled.